Avoid doing bad things on inode type changes during revalidation.
When Coda discovers an inconsistent object, it turns it into a symlink.
However we can't just follow this change in the kernel on an existing
file or directory inode that may still have references.

Remove the inconsistent inode from the inode hash so that we allocate a new
one for the symlink object.